chore: rename new package to commands
